requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
asyncpg>=0.24.0
pgvector>=0.2.0
python-dotenv>=0.19.0
pydantic>=1.8.0
numpy>=1.21.0
sentence-transformers>=2.2.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
huggingface-hub==0.16.4
prometheus-client==0.19.0
gunicorn==21.2.0
cryptography==42.0.2
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0 